requirements.txt
openai
pdfplumber
openpyxl
python-dotenv
